requirements.txt
peewee~=3.16.0
Flask~=2.2.3
pydantic~=1.10.7
Flask-Pydantic~=0.11.0
PyJWT~=2.6.0
flask_cors~=3.0.10